+++ src/Selections/ScanSelection.h
@@ -27,8 +27,8 @@ private:
     GlobalContext& context = GlobalContext::getInstance();
     CardputerView& display;
     CardputerInput& input;
-    uint8_t selectionIndex;
-    int8_t lastIndex;
+    uint16_t selectionIndex;
+    int16_t lastIndex;
 };
 
 }
